feat: basic implementation of JSON:API plugin(reponse only)
This is the start of a fully compliant JSON API implementation. This
plugin is now capable of transforming a response body into a valid JSON
API response.

Check out this
[gist](https://gist.github.com/eliias/ed80dc776bf505dca970) for the feature status of this commit.

BREAKING CHANGE: will break if you have used the *jsonapi* plugin

The previous JSON API implementation was just a basic wrapper around a
list of plain objects or an object itself without any support of JSON
API semantics. If you want to use JSON API, you have to change the code
of your clients.