feat: full relationship management
This is a major change in comparison to the previous version.
Relationships are now defined in a more concise way. Beside that, all
methods do support One-To-Many and Many-To-One relationships now. Even
create and delete are now supported.

Test coverage for all REST tests has been improved and are more strict
overall. Tests are also grouped and should not produce races anymore.

BREAKING CHANGE: The setup of routes has changed. The methods
`hasMany`and `belongsTo` have been introduced. Relationships must be
defined with the `relationship` property within the REST plugin setup.